Remove superfluous viewFromColumn from Matrix
Before the change, a matrix view was inheriting a portion of data
matching whole rows of the initial matrix, so there was the need to
keep a record of the column offset.

It was intuitive and easy to debug, but the Matrix type needed to
store and include in offset calculations an additional property
(viewFromColumn).

This change eliminates the need for that property. The column offset
is already considered when slicing the data for a new view. In this way,
data[0] always corresponds to the element at 0x0 (granted that size
is at least 1x1).

Just out of consistency, and slightly easier debuggability, the right
side of a view's data is trimmed too, when appropriate according
to the view's parameters. In this way, the excess tail from the last
row of the parent matrix is hidden from the view's data.